Add endpoint to get business by user id

+++ server/controller/business.js
@@ -1,5 +1,7 @@
 import models from '../models/index';
 import errorMessage from '../middlewares/error-message';
+import errorHandler from '../middlewares/error-handler';
+import businessValidator from '../validation/business';
 
 const Businesses = models.Business;
 /**
@@ -24,6 +26,21 @@ class BusinessMethods {
       category
     } = req.body;
 
+    const conflict = {};
+
+    const { errors, isValid } = businessValidator(req.body);
+
+    if (!isValid) {
+      return res.status(400).json(errors);
+    }
+
+    if (req.decoded.username === 'admin') {
+      return res.status(403).json({
+        message: 'An admin can not register a business',
+        error: true
+      });
+    }
+
     return Businesses.create({
       name: name.toLowerCase(),
       email: email.toLowerCase(),
@@ -37,10 +54,13 @@ class BusinessMethods {
         error: false,
         business
       }))
-      .catch(error => res.status(409).json({
-        message: error.errors[0].message,
-        error: true
-      }));
+      .catch((error) => {
+        if (error.errors[0].path === 'name') {
+          conflict.businessname = 'Business with name is already registered';
+        }
+
+        return res.status(409).json(conflict);
+      });
   }
   /**
     *
@@ -58,10 +78,7 @@ class BusinessMethods {
             message: 'No business registered yet'
           });
         }
-        return res.status(200).json({
-          businesses,
-          error: false
-        });
+        return res.status(200).json({ businesses });
       });
   }
   /**
@@ -79,9 +96,36 @@ class BusinessMethods {
         }
 
         return res.status(200).json({
-          business,
+          business: [business],
           error: false
         });
+      })
+      .catch(error => res.status(500).json({
+        message: error.message,
+        help: 'No space, only an integer is allowed',
+        error: true
+      }));
+  }
+  /**
+          *
+          *@param {any} req - request value
+          *@param {any} res - response value
+          *@return {json} response object gotten
+          *@memberof BusinessMethods
+        */
+  static getBusinessByUserId(req, res) {
+    const { userId } = req.params;
+    return Businesses.findAll({
+      where: {
+        userId
+      }
+    })
+      .then((businesses) => {
+        if (businesses === null) {
+          return errorMessage(res);
+        }
+
+        return res.status(200).json({ businesses });
       })
       .catch(error => res.status(500).json({
         message: error.message,
@@ -97,9 +141,18 @@ class BusinessMethods {
     *@memberof BusinessMethods
   */
   static updateBusiness(req, res) {
+    const conflict = {};
+
     const businessId = parseInt(req.params.businessId, 10);
     return Businesses.findById(businessId)
       .then((business) => {
+        if (business.userId !== req.decoded.userId) {
+          return res.status(403).json({
+            message: 'Forbidden, you do not have access to modify this business',
+            error: true
+          });
+        }
+
         business.update({
           name: req.body.name || business.name,
           email: req.body.email || business.email,
@@ -111,7 +164,14 @@ class BusinessMethods {
             message: 'Business updated successfully',
             error: false,
             business
-          }));
+          }))
+          .catch((error) => {
+            if (error.errors[0].path === 'name') {
+              conflict.businessname = 'Business with name is already registered';
+            }
+
+            return res.status(409).json(conflict);
+          });
       });
   }
   /**
@@ -123,15 +183,27 @@ class BusinessMethods {
   */
   static deleteBusiness(req, res) {
     const businessId = parseInt(req.params.businessId, 10);
-    return Businesses.destroy({
+    return Businesses.findOne({
       where: {
         id: businessId
       }
-    })
-      .then(() => res.status(200).json({
-        message: 'Business deleted successfully',
-        error: false
-      }));
+    }).then((business) => {
+      if (req.decoded.username === 'admin' || business.userId === req.decoded.userId) {
+        return Businesses.destroy({
+          where: {
+            id: businessId
+          }
+        })
+          .then(() => res.status(200).json({
+            message: 'Business deleted successfully',
+            error: false
+          }));
+      }
+      return res.status(403).json({
+        message: 'Forbidden, you do not have access to modify this business',
+        error: true
+      });
+    });
   }
 }
